@@ -9,29 +9,31 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@auth/firebase-adapter": "^2.1.0",
     "@hookform/resolvers": "^3.3.4",
-    "@nextui-org/react": "^2.2.9",
+    "@nextui-org/react": "^2.4.1",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-slot": "^1.0.2",
-    "@serwist/next": "^8.4.4",
-    "@serwist/precaching": "^8.4.4",
-    "@serwist/sw": "^8.4.4",
+    "@serwist/next": "^9.0.2",
     "apexcharts": "^3.46.0",
     "axios": "^1.6.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "cmdk": "^0.2.1",
     "date-fns": "^3.3.1",
+    "eslint-config-next": "^14.2.3",
+    "firebase": "^10.12.1",
+    "firebase-admin": "^12.1.1",
     "framer-motion": "^11.0.5",
     "lucide-react": "^0.331.0",
     "luxon": "^3.4.4",
-    "next": "^14.1.0",
-    "next-auth": "^4.24.5",
-    "react": "^18.2.0",
+    "next": "^14.2.3",
+    "next-auth": "^4.24.7",
+    "react": "^18.3.1",
     "react-apexcharts": "^1.4.1",
     "react-datepicker": "^6.1.0",
     "react-day-picker": "^8.10.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.50.1",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.0.1",
@@ -50,6 +52,7 @@
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "postcss": "^8",
+    "serwist": "^9.0.2",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
   }
