@@ -1,13 +1,22 @@
-/** @type {import('next').NextConfig} */
+const { PHASE_DEVELOPMENT_SERVER, PHASE_PRODUCTION_BUILD } = require("next/constants");
 
-const withSerwist = require("@serwist/next").default({
-	disable: process.env.NODE_ENV === "development",
-	swSrc: "src/sw.ts",
-	swDest: "public/sw.js",
-	register: true,
-});
-
-module.exports = withSerwist({
+const nextConfig = {
 	swcMinify: true,
 	output: "standalone",
-});
+};
+
+/** @type {(phase: string, defaultConfig: import("next").NextConfig) => Promise<import("next").NextConfig>} */
+module.exports = async (phase) => {
+	if (phase === PHASE_DEVELOPMENT_SERVER || phase === PHASE_PRODUCTION_BUILD) {
+		const withSerwist = (await import("@serwist/next")).default({
+			cacheOnNavigation: true,
+			disable: process.env.NODE_ENV === "development",
+			swSrc: "src/sw.ts",
+			swDest: "public/sw.js",
+			register: true,
+		});
+		return withSerwist(nextConfig);
+	}
+
+	return nextConfig;
+};
