@@ -3,6 +3,7 @@ import { Control, Controller, FieldError } from "react-hook-form";
 import { Chip, Select as NextSelect, SelectItem as NextSelectItem, SelectedItems } from "@nextui-org/react";
 import { IOption } from "@/types/Forms";
 import { capitalize } from "@/utils";
+import { Key } from "@react-types/shared";
 
 interface AppSelectProps {
 	name?: string;
@@ -50,50 +51,61 @@ const AppSelect = ({
 		<Controller
 			control={control}
 			name={name}
-			render={({ field: { onChange: onControlledChange, value: controlledValue } }) => (
-				<NextSelect
-					label={label}
-					value={controlledValue}
-					onChange={(val) => {
-						onControlledChange(val);
-						onSelectAction && onSelectAction();
-					}}
-					description={helperText}
-					isInvalid={!!error}
-					errorMessage={error?.message}
-					size="md"
-					variant="bordered"
-					labelPlacement="outside"
-					placeholder={placeholder}
-					items={options}
-					classNames={{
-						label: "text-sm font-medium text-secondary",
-						base: baseClassName,
-					}}
-					renderValue={(items) =>
-						customRender ? (
-							customRender(items)
-						) : (
-							<div className="flex flex-wrap gap-2">
-								{items.map((item) => (
-									<Chip color="primary" key={item.key} className="text-[12px]" size="sm">
-										{item.key as string}
-									</Chip>
-								))}
-							</div>
-						)
-					}>
-					{(item) => {
-						const opt = getOptionItem(item);
+			render={({ field: { onChange: onControlledChange, value: controlledValue } }) => {
+				return (
+					<NextSelect
+						label={label}
+						// value={controlledValue}
+						// onChange={(val) => {
+						// 	console.log(val);
+						// 	onControlledChange(val);
+						// 	onSelectAction && onSelectAction();
+						// }}
+						description={helperText}
+						isInvalid={!!error}
+						errorMessage={error?.message}
+						size="md"
+						variant="bordered"
+						labelPlacement="outside"
+						placeholder={placeholder}
+						items={options as Iterable<IOption>}
+						classNames={{
+							label: "text-sm font-medium text-secondary",
+							base: baseClassName,
+						}}
+						selectedKeys={[controlledValue ?? ""]}
+						onSelectionChange={(val: Set<Key>) => {
+							const selectedItem = Array.from(val)?.[0];
+							onControlledChange(selectedItem);
+							onSelectAction && onSelectAction();
+						}}
+						renderValue={(items) =>
+							customRender ? (
+								customRender(items)
+							) : (
+								<div className="flex flex-wrap gap-2">
+									{items.map((item) => {
+										return (
+											<Chip color="primary" key={item.key} className="text-[12px]" size="sm">
+												{item.key as string}
+											</Chip>
+										);
+									})}
+								</div>
+							)
+						}>
+						{(item) => {
+							const opt = getOptionItem(item);
 
-						return (
-							<NextSelectItem key={opt?.value ?? "new-key"} value={opt.value}>
-								{opt.label}
-							</NextSelectItem>
-						);
-					}}
-				</NextSelect>
-			)}
+							return (
+								<NextSelectItem key={opt?.value ?? "new-key"} value={opt.value} textValue={opt?.value}>
+									{opt.label}
+								</NextSelectItem>
+							);
+						}}
+					</NextSelect>
+				);
+			}}
 		/>
 	) : (
 		<NextSelect
@@ -118,7 +130,7 @@ const AppSelect = ({
 				label: "text-sm font-medium text-secondary",
 				base: baseClassName,
 			}}
-			items={options}
+			items={options as Iterable<IOption>}
 			renderValue={(items) =>
 				customRender ? (
 					customRender(items)
