@@ -48,7 +48,7 @@ export const appNavbarMenuItems = [
 				<AppNavLinkItem title="Loan Requests" icon={<HiBriefcase className="text-primary" size={18} />} href="green-finance/loans" />
 			</>
 		)}
-		show={false}
+		show={true}
 	/>,
 	<AppNavLinkSection
 		title="Company"
