@@ -12,9 +12,10 @@ interface StationaryCombustionConfirmModalProps {
 	values?: Omit<IScopeOneFuels, "id" | "createdAt" | "updatedAt"> & { date: string | Date };
 	onConfirm?: VoidFunction;
 	isSaving?: boolean;
+	actionType?: "create" | "update";
 }
 
-const StationaryCombustionConfirmModal = ({ isOpen, setIsOpen, values, onConfirm, isSaving }: StationaryCombustionConfirmModalProps) => {
+const StationaryCombustionConfirmModal = ({ isOpen, setIsOpen, values, onConfirm, isSaving, actionType = "create" }: StationaryCombustionConfirmModalProps) => {
 	return (
 		<>
 			<Modal isOpen={isOpen} onOpenChange={setIsOpen} size="4xl" scrollBehavior="outside">
@@ -48,10 +49,10 @@ const StationaryCombustionConfirmModal = ({ isOpen, setIsOpen, values, onConfirm
 							</ModalBody>
 							<ModalFooter>
 								<Button color="primary" variant="bordered" startContent={<FiEdit3 />} onPress={onClose}>
-									Edit
+									{actionType === "create" ? "Edit" : "Cancel"}
 								</Button>
 								<Button color="primary" startContent={<CheckIcon className="w-4 h-4" />} onPress={onConfirm} isLoading={isSaving} isDisabled={isSaving}>
-									Confirm
+									{actionType === "update" ? "Update" : "Confirm"}
 								</Button>
 							</ModalFooter>
 						</>
