@@ -1,21 +1,66 @@
-import { defaultCache } from "@serwist/next/browser";
-import type { PrecacheEntry } from "@serwist/precaching";
-import { installSerwist } from "@serwist/sw";
+import { ExpirationPlugin, NetworkFirst, PrecacheEntry, Serwist, SerwistGlobalConfig } from "serwist";
+import { PAGES_CACHE_NAME, defaultCache } from "@serwist/next/worker";
 
-declare const self: ServiceWorkerGlobalScope & {
-	// Change this attribute's name to your `injectionPoint`.
-	// `injectionPoint` is an InjectManifest option.
-	// See https://serwist.pages.dev/docs/build/inject-manifest/configuring
-	__SW_MANIFEST: (PrecacheEntry | string)[] | undefined;
-};
+// This declares the value of `injectionPoint` to TypeScript.
+// `injectionPoint` is the string that will be replaced by the
+// actual precache manifest. By default, this string is set to
+// `"self.__SW_MANIFEST"`.
+declare global {
+	interface WorkerGlobalScope extends SerwistGlobalConfig {
+		__SW_MANIFEST: (PrecacheEntry | string)[] | undefined;
+	}
+}
 
-installSerwist({
+declare const self: ServiceWorkerGlobalScope;
+
+const serwist = new Serwist({
 	precacheEntries: self.__SW_MANIFEST,
 	skipWaiting: true,
 	clientsClaim: true,
 	navigationPreload: true,
-	runtimeCaching: defaultCache,
+	runtimeCaching: [
+		{
+			matcher: ({ request, url: { pathname }, sameOrigin }) => request.headers.get("RSC") === "1" && request.headers.get("Next-Router-Prefetch") === "1" && sameOrigin && !pathname.startsWith("/api/"),
+			// NEW: an initialized instance.
+			handler: new NetworkFirst({
+				cacheName: PAGES_CACHE_NAME.rscPrefetch,
+				plugins: [
+					new ExpirationPlugin({
+						maxEntries: 32,
+						maxAgeSeconds: 7 * 24 * 60 * 60, // 7 days
+					}),
+				],
+			}),
+		},
+		{
+			matcher: ({ request, url: { pathname }, sameOrigin }) => request.headers.get("RSC") === "1" && sameOrigin && !pathname.startsWith("/api/"),
+			// NEW: an initialized instance.
+			handler: new NetworkFirst({
+				cacheName: PAGES_CACHE_NAME.rsc,
+				plugins: [
+					new ExpirationPlugin({
+						maxEntries: 32,
+						maxAgeSeconds: 7 * 24 * 60 * 60, // 7 days
+					}),
+				],
+			}),
+		},
+		{
+			matcher: ({ request, url: { pathname }, sameOrigin }) => request.headers.get("Content-Type")?.includes("text/html") && sameOrigin && !pathname.startsWith("/api/"),
+			// NEW: an initialized instance.
+			handler: new NetworkFirst({
+				cacheName: PAGES_CACHE_NAME.html,
+				plugins: [
+					new ExpirationPlugin({
+						maxEntries: 32,
+						maxAgeSeconds: 7 * 24 * 60 * 60, // 7 days
+					}),
+				],
+			}),
+		},
+
+		...defaultCache,
+	],
 });
 
-// TODO: Must implement this after full migration to app folder.
-// * https://serwist.pages.dev/docs/next/getting-started
+serwist.addEventListeners();
