@@ -40,9 +40,9 @@ export const enum IApiEndpoint {
 	GET_COMPANY = "company/profile",
 	ADMIN_CREATE_COMPANY = "company/admin/create",
 	ADMIN_UPDATE_USER_WITH_COMPANY = "company/admin/add-user-to-company",
-	GET_COMPANY_BRANCHES="company/branches/company",
-	CREATE_NEW_COMPANY_BRANCH="company/branches/new",
-	DELETE_COMPANY_BRANCH="company/branch/remove",
+	GET_COMPANY_BRANCHES = "company/branches/company",
+	CREATE_NEW_COMPANY_BRANCH = "company/branches/new",
+	DELETE_COMPANY_BRANCH = "company/branch/remove",
 
 	WAITLIST = "waitlist",
 
@@ -81,6 +81,25 @@ export const enum IApiEndpoint {
 	GET_INVITE_INFO = "invites/info/code",
 	REJECT_INVITE = "invites/company/reject",
 	REVOKE_INVITE = "invites/revoke",
+
+	// GET SCOPE TWO DATA ITEMS
+	GET_SCOPE_TWO_ELECTRICTY_DATA = "scope-two/get/electricity-data",
+
+	// GET SCOPE ONE DATA ITEMS
+	GET_SCOPE_ONE_VEHICLE_EMISSIONS_DATA = "scope-one/get/vehicles-emissions",
+	GET_SCOPE_ONE_PROCESS_EMISSIONS_DATA = "scope-one/get/process-emissions",
+	GET_SCOPE_ONE_FUGITIVE_EMISSIONS_DATA = "scope-one/get/fugitive-emissions",
+	GET_SCOPE_ONE_FUELS_EMISSIONS_DATA = "scope-one/get/fuel-emissions",
+
+	// UPDATE ACCOUNTING DATA
+	UPDATE_SCOPE_TWO_ELECTRICITY_DATA = "scope-two/update/electricity-data",
+	UPDATE_SCOPE_TWO_HEAT_AND_COOLING_DATA = "scope-two/update/heat-and-cooling-data",
+
+	// UPDATE SCOPE ONE ACCOUNTING DATA
+	UPDATE_SCOPE_ONE_VEHICLE_EMISSIONS_DATA = "scope-one/update/vehicles-emissions",
+	UPDATE_SCOPE_ONE_PROCESSING_EMISSIONS_DATA = "scope-one/update/process-emissions",
+	UPDATE_SCOPE_ONE_FUGITIVE_EMISSIONS_DATA = "scope-one/update/fugitive-emissions",
+	UPDATE_SCOPE_ONE_FUEL_EMISSIONS_DATA = "scope-one/update/fuel-emissions",
 }
 
 export interface IMethodParams {
